ревакторит, убирает из компонента SearchBar типы и формулы
