+++ src/types/search-bar.ts
@@ -1,0 +1,7 @@
+type SearchItem = {
+  name: string;
+  vendorCode: string;
+  id: number;
+}
+
+export type SearchItems = SearchItem[];
